refactor(arns): move splitPrimaryName into src/utils per review
Relocate the ArNS name splitter from `src/solana/name.ts` to the shared
`src/utils/` directory (`src/utils/arns.ts`), matching the repo convention
where pure helpers live in utils with a co-located test and a barrel export.
Addresses review feedback that this belongs in the existing utils file.

- add `src/utils/arns.ts` + co-located `src/utils/arns.test.ts` (the
  splitPrimaryName cases move here from `primary-name.test.ts`)
- export it from `src/utils/index.js`
- repoint importers (`io-readable`, `io-writeable`, the two tests) to
  `../utils/arns.js`
- delete `src/solana/name.ts`

No behavior change. Note this exposes `splitPrimaryName` publicly, since
`utils/index.js` is re-exported from the package entry.
